requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
PyYAML==6.0.1
openai==1.3.5
